tfjson: Add back all the overloads
It was a mistake to remove them before, just causing extra caller casts.

+++ src/ocpp/Connection.cpp
@@ -193,7 +193,7 @@ void OcppConnection::handleCallError(uint64_t uid, CallErrorCode code, const cha
 
 static size_t buildCallError(TFJsonSerializer &json, const char *uid, CallErrorCode code, const char *desc) {
     json.addArray();
-    json.add((int64_t)OcppRpcMessageType::CALLERROR);
+    json.add((int32_t)OcppRpcMessageType::CALLERROR);
     json.add(uid);
     json.add(CallErrorCodeStrings[(size_t)code]);
     json.add(desc);
